refactor: simplify naming
