+++ apps/api-journeys/src/app/modules/team/team.module.ts
@@ -1,13 +1,13 @@
 import { Global, Module } from '@nestjs/common'
 import { DatabaseModule } from '@core/nest/database/DatabaseModule'
-import { AuthModule } from '@core/nest/common/AuthModule'
+import { CaslAuthModule } from '@core/nest/common/CaslAuthModule'
 import { PrismaService } from '../../lib/prisma.service'
 import { AppCaslFactory } from '../../lib/casl/caslFactory'
 import { TeamResolver } from './team.resolver'
 
 @Global()
 @Module({
-  imports: [DatabaseModule, AuthModule.register(AppCaslFactory)],
+  imports: [DatabaseModule, CaslAuthModule.register(AppCaslFactory)],
   providers: [TeamResolver, PrismaService],
   exports: []
 })
