--- libs/nest/common/src/lib/AuthModule/guards/roles.guard.spec.ts
@@ -1,22 +1,0 @@
-import { RolesGuard, rbacLogic } from './roles.guard'
-
-describe('rbac-logic', () => {
-  it('should pass when user has one of the defined roles', () => {
-    expect(rbacLogic(['Editor'], ['Admin', 'Editor'])).toEqual(true)
-  })
-
-  it('should pass if user has Super role but not explicitly defined', () => {
-    expect(rbacLogic(['Super'], ['Admin', 'Editor'])).toEqual(true)
-  })
-
-  it('should fail if user does not contain any of the defined roles', () => {
-    expect(rbacLogic(['Registered'], ['Admin', 'Editor'])).toEqual(false)
-  })
-
-  it('should throw if called with new', () => {
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    expect(() => new (RolesGuard as any)()).toThrow(
-      'RolesGuard cannot be instantiated directly. Use RolesGuard() instead.'
-    )
-  })
-})
