--- libs/nest/common/src/lib/AuthModule/index.ts
@@ -1,8 +1,0 @@
-export { AllowAnonymous } from './decorators/allowAnonymous.decorator'
-export { CaslAbility } from './decorators/caslAbility.decorator'
-export { CaslAccessible } from './decorators/caslAccessible.decorator'
-export { CaslPolicy } from './decorators/caslPolicy.decorator'
-export { CaslGuard } from './guards/casl.guard'
-export { RolesGuard } from './guards/roles.guard'
-export { CaslFactory } from './casl.factory'
-export { AuthModule } from './auth.module'
