+++ libs/nest/common/src/lib/CaslAuthModule/caslGuard.ts
@@ -2,12 +2,8 @@ import { CanActivate, ExecutionContext, Injectable } from '@nestjs/common'
 import { Reflector } from '@nestjs/core'
 import { GqlExecutionContext } from '@nestjs/graphql'
 
-import { CaslFactory } from '../casl.factory'
-import { ALLOW_ANONYMOUS_KEY } from '../decorators/allowAnonymous.decorator'
-import {
-  CASL_POLICY_KEY,
-  CaslPolicyHandler
-} from '../decorators/caslPolicy.decorator'
+import { CaslFactory } from './caslFactory'
+import { CASL_POLICY_KEY, CaslPolicyHandler } from './decorators/caslPolicy'
 
 /**
  * Guard that is used in conjunction with `CaslAbility`, `CaslAccessible` and `CaslPolicy` decorators.
@@ -35,18 +31,6 @@ export class CaslGuard implements CanActivate {
   ) {}
 
   async canActivate(context: ExecutionContext): Promise<boolean> {
-    const allowAnonymousHandler = this.reflector.get<boolean | undefined>(
-      ALLOW_ANONYMOUS_KEY,
-      context.getHandler()
-    )
-    if (allowAnonymousHandler === true) return true
-
-    const allowAnonymousClass = this.reflector.get<boolean | undefined>(
-      ALLOW_ANONYMOUS_KEY,
-      context.getClass()
-    )
-    if (allowAnonymousClass === true) return true
-
     const req = GqlExecutionContext.create(context).getContext().req
 
     if (req.userId == null) return false
